pyproject.toml
[project]
name = "literature-review-agent"
version = "0.1.0" 
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langfuse (>=3.0.7,<4.0.0)",
    "langgraph (>=0.5.0,<0.6.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "arxiv (>=2.2.0,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "feedparser (>=6.0.11,<7.0.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "pinecone (>=7.3.0,<8.0.0)"
    
]

[tool.poetry]              
packages = [
  { include = "agents",    from = "src" },
  { include = "api",       from = "src" },
  { include = "core",      from = "src" },
  { include = "models",    from = "src" },
  { include = "services",  from = "src" },
  { include = "workflows", from = "src" },
  { include = "utils",     from = "src" }
]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.16.1"
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
